+++ src/cookiebanner.js
@@ -225,7 +225,12 @@ THE SOFTWARE.
                 }
             }
             return null;
-        }
+        },
+
+        is_numeric: function(val) {
+            // eslint-disable-next-line eqeqeq
+            return (parseFloat(val) == val); // want type coercion here!
+        },
 
     };
 
@@ -315,10 +320,16 @@ THE SOFTWARE.
             this.options.mask = Utils.str2bool(this.options.mask);
             this.options.closePrecedes = Utils.str2bool(this.options.closePrecedes);
 
-            // check for a possible global callback specified as a string
+			// check for a possible global callback specified as a string
             if ('string' === typeof this.options.expires) {
                 if ('function' === typeof context[this.options.expires]) {
                     this.options.expires = context[this.options.expires];
+                } else {
+                    // No global callback under that name, check if it's a
+                    // number/numeric and use as duration in seconds.
+                    if (Utils.is_numeric(this.options.expires)) {
+                        this.options.expires = Number(parseFloat(this.options.expires));
+                    }
                 }
             }